desafio_rpa/requirements.txt
loguru>=0.7.2
selenium>=4.16.0
openpyxl>=3.1.2
pathlib>=1.0.1
python-dotenv>=1.0.0